+++ Engine_SOURCE/Engine_SOURCE.vcxproj.filters
@@ -91,14 +91,11 @@
     <ClCompile Include="yaTime.cpp">
       <Filter>Engine\Time</Filter>
     </ClCompile>
-    <ClCompile Include="MonsterBase.cpp">
+    <ClCompile Include="yaMonsterBase.cpp">
       <Filter>Contents\Monster\Base</Filter>
     </ClCompile>
     <ClCompile Include="yaApplication.cpp">
       <Filter>Engine\High Level InterFace</Filter>
-    </ClCompile>
-    <ClCompile Include="yaBaseRenderer.cpp">
-      <Filter>Engine\Resource\Renderer</Filter>
     </ClCompile>
     <ClCompile Include="yaGameObject.cpp">
       <Filter>Engine\GameObject</Filter>
@@ -111,9 +108,6 @@
     </ClCompile>
     <ClCompile Include="yaStructedBuffer.cpp">
       <Filter>Engine\Graphics\Buffer</Filter>
-    </ClCompile>
-    <ClCompile Include="yaRenderer.cpp">
-      <Filter>Engine\Resource\Renderer</Filter>
     </ClCompile>
     <ClCompile Include="yaScene.cpp">
       <Filter>Engine\Scene\Base</Filter>
@@ -202,7 +196,14 @@
     <ClCompile Include="yaMusketeerman_Shooting.cpp">
       <Filter>Contents\Monster\Attack</Filter>
     </ClCompile>
-
+    <ClCompile Include="yaBaseRenderer.cpp" />
+    <ClCompile Include="yaLogBridgeScript.cpp" />
+    <ClCompile Include="yaMeshData.cpp" />
+    <ClCompile Include="yaPlayerActionScript.cpp" />
+    <ClCompile Include="yaRenderer.cpp" />
+    <ClCompile Include="yaSwordsman.cpp">
+      <Filter>Contents\Monster\Swordsman</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="yaSekiroEnums.h" />
@@ -302,14 +303,11 @@
     <ClInclude Include="yaTime.h">
       <Filter>Engine\Time</Filter>
     </ClInclude>
-    <ClInclude Include="MonsterBase.h">
+    <ClInclude Include="yaMonsterBase.h">
       <Filter>Contents\Monster\Base</Filter>
     </ClInclude>
     <ClInclude Include="yaApplication.h">
       <Filter>Engine\High Level InterFace</Filter>
-    </ClInclude>
-    <ClInclude Include="yaBaseRenderer.h">
-      <Filter>Engine\Resource\Renderer</Filter>
     </ClInclude>
     <ClInclude Include="yaGameObject.h">
       <Filter>Engine\GameObject</Filter>
@@ -328,9 +326,6 @@
     </ClInclude>
     <ClInclude Include="yaGraphics.h">
       <Filter>Engine\Graphics</Filter>
-    </ClInclude>
-    <ClInclude Include="yaRenderer.h">
-      <Filter>Engine\Resource\Renderer</Filter>
     </ClInclude>
     <ClInclude Include="yaScene.h">
       <Filter>Engine\Scene\Base</Filter>
@@ -418,12 +413,18 @@
     </ClInclude>
     <ClInclude Include="yaMusketeerman_Shooting.h">
       <Filter>Contents\Monster\Attack</Filter>
-	</ClInclude>
+    </ClInclude>
     <ClInclude Include="yaPlayerActionScript.h">
       <Filter>Contents\Player</Filter>
-	</ClInclude>
+    </ClInclude>
     <ClInclude Include="yaLogBridgeScript.h">
       <Filter>Contents\Ground</Filter>
+    </ClInclude>
+    <ClInclude Include="yaBaseRenderer.h" />
+    <ClInclude Include="yaMeshData.h" />
+    <ClInclude Include="yaRenderer.h" />
+    <ClInclude Include="yaSwordsman.h">
+      <Filter>Contents\Monster\Swordsman</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -585,6 +586,9 @@
     </Filter>
     <Filter Include="Contents\Monster\Musketeerman">
       <UniqueIdentifier>{351c027e-06f9-4c40-85e6-15e96efeb633}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Contents\Monster\Swordsman">
+      <UniqueIdentifier>{dacc8e74-def5-445a-a26c-2ec4d855704d}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
