+++ Engine_SOURCE/yaScene.h
@@ -26,8 +26,9 @@ namespace ya
 		Layer& GetLayer(eLayerType type) { return mLayers[(UINT)type]; }
 		std::vector<GameObject*> GetDontDestroyGameObjects();
 		const std::vector<GameObject*>& GetGameObjects(const eLayerType type);
+		
 
-
+		
 	private:
 		std::vector<Layer> mLayers;
 		eSceneType mType;
