+++ Engine_SOURCE/yaSpearman.cpp
@@ -19,6 +19,19 @@ namespace ya
 		faceRenderer->SetMaterial(Resources::Find<Material>(L"BasicMaterial"), 0);
 
 
+		{
+			Transform* tr = GetComponent<Transform>();
+			GameObject* face = object::Instantiate<GameObject>(eLayerType::Player, tr);
+			face->SetName(L"TEST");
+			TEST = face->GetComponent<Transform>();
+			TEST->SetPosition(Vector3(100.0f, 100.0f, 100.0f));
+			TEST->SetScale(Vector3(1.3f, 1.3f, 0.3f));
+			MeshRenderer* faceRenderer = face->AddComponent<MeshRenderer>();
+			faceRenderer->SetMesh(Resources::Find<Mesh>(L"CubeMesh"));
+			faceRenderer->SetMaterial(Resources::Find<Material>(L"BasicMaterial"), 0);
+		}
+
+
 		CreateMonsterState();
 		SetSituation(enums::eSituation::None);
 
@@ -40,17 +53,58 @@ namespace ya
 		Vec3 playerPos = GetPlayerPos();
 		Vec3 monsterPos = GetComponent<Transform>()->GetPosition();
 
-
+		
 
 		switch (GetSituation())
 		{
 		case ya::enums::eSituation::None:
 		{
-			mTime += Time::DeltaTime();
-			if (mTime >= 3.0f)
+			//인살 띄우기
+			if (IsDeathBlow())
 			{
-				SetSituation(enums::eSituation::Battle);
-				mTime = 0.f;
+				TEST->SetPosition(Vector3(0.0f, 2.0f, 0.0f));
+
+				if (IsStartBlow())
+				{
+					if (!NavigationPlayer(3.0f))
+					{
+						TEST->SetPosition(Vector3(100.0f, 100.0f, 100.0f));
+					}
+				}
+			}
+			//없애기
+			else
+			{
+				TEST->SetPosition(Vector3(100.0f, 100.0f, 100.0f));
+			}
+
+			if (IsPlayerFieldview())
+			{
+				if (NavigationPlayer(15.0f))
+				{
+					//배틀로 상태 변경
+					SetStartBlow(false);
+					if (GetSituation() != enums::eSituation::Groggy)
+					{
+						SetDeathBlow(false);
+					}
+					SetSituation(enums::eSituation::Battle);
+				}
+			}
+			else
+			{
+				if (!IsPlayerFront())
+				{
+					if (NavigationPlayer(3.0f))
+					{
+						int a = 0;
+						if (IsStartBlow())
+							SetDeathBlow(true);
+					}
+
+
+				}
+
 			}
 		}
 			break;
@@ -62,9 +116,9 @@ namespace ya
 			break;
 		case ya::enums::eSituation::Battle:
 		{
-			//int random = RandomNumber(1, 3);
+			int random = RandomNumber(1, 3);
 
-			int random = 3;
+			
 
 			if (random == 1)
 			{				
@@ -82,7 +136,7 @@ namespace ya
 			}
 			else if (random == 2)
 			{
-				SetSituation(enums::eSituation::Defense);		
+				SetSituation(enums::eSituation::Defense);
 				TurnToPlayer();
 			}
 			else if (random == 3)
@@ -136,7 +190,6 @@ namespace ya
 				SetSituation(enums::eSituation::None);
 				mTime = 0.f;				
 			}
-
 		}
 			break;
 		case ya::enums::eSituation::Attack:
@@ -154,16 +207,20 @@ namespace ya
 				Attack_sting();
 				SetSituation(enums::eSituation::None);
 			}
-			
-			
 		}
 			break;
 		case ya::enums::eSituation::Sit:
+			break;
+		case ya::enums::eSituation::Groggy:
+		{			
+
+		}
 			break;
 		case ya::enums::eSituation::Death:
 			break;
 		case ya::enums::eSituation::End:
 			break;
+
 		}
