+++ Engine_SOURCE/yaTitleScene.cpp
@@ -29,6 +29,8 @@
 
 #include "yaSpearman.h"
 #include "yaMusketeerman.h"
+#include "yaSwordsman.h"
+
 #include "yaMonsterScript.h"
 #include "yaFbxLoader.h"
 #include "yaMeshData.h"
@@ -54,7 +56,7 @@ namespace ya
 		CameraScript* camScript = cameraObj->AddComponent<CameraScript>();
 		mainCamera = cameraComp;
 
-
+		
 		Player* player = object::Instantiate<Player>(eLayerType::Player);
 		player->GetComponent<Transform>()->SetPosition(Vector3(-13.0f, 0.0f, -10.0f));
 		player->GetComponent<Transform>()->SetScale(Vector3(5.0f, 5.0f, 5.0f));
@@ -75,8 +77,8 @@ namespace ya
 			spearman->SetName(L"Spearman");
 			spearman->SetPlayerObject(player);
 			MeshRenderer* spearmanmr = spearman->AddComponent<MeshRenderer>();
-			spearmanmr->SetMaterial(Resources::Find<Material>(L"BasicMaterial"));
 			spearmanmr->SetMesh(Resources::Find<Mesh>(L"CubeMesh"));
+			spearmanmr->SetMaterial(Resources::Find<Material>(L"BasicMaterial"), 0);
 			Collider2D* spearmancol = spearman->AddComponent <Collider2D>();
 			spearmancol->SetType(eColliderType::Box);
 			spearmancol->SetSize(Vector3(1.0, 2.0f, 2.0f));
@@ -88,31 +90,50 @@ namespace ya
 		}
 
 
-		{
-			Musketeerman* musketeerman = object::Instantiate<Musketeerman>(eLayerType::Monster);
-			musketeerman->GetComponent<Transform>()->SetPosition(Vector3(-5.0f, 0.0f, 15.0f));
-			musketeerman->GetComponent<Transform>()->SetScale(Vector3(5.0f, 5.0f, 5.0f));
-			musketeerman->SetName(L"Musketeerman");
-			musketeerman->SetPlayerObject(player);
-			MeshRenderer* musketeermanmr = musketeerman->AddComponent<MeshRenderer>();
-			musketeermanmr->SetMesh(Resources::Find<Mesh>(L"CubeMesh"));
-			musketeermanmr->SetMaterial(Resources::Find<Material>(L"BasicMaterial"), 0);
-			Collider2D* musketeermancol = musketeerman->AddComponent <Collider2D>();
-			musketeermancol->SetType(eColliderType::Box);
-			musketeermancol->SetSize(Vector3(1.0, 2.0f, 2.0f));
-			Rigidbody* spearmanRigidbody = musketeerman->AddComponent<Rigidbody>();
-			spearmanRigidbody->SetGround(false);
+		//{
+		//	Musketeerman* musketeerman = object::Instantiate<Musketeerman>(eLayerType::Monster);
+		//	musketeerman->GetComponent<Transform>()->SetPosition(Vector3(-5.0f, 0.0f, 15.0f));
+		//	musketeerman->GetComponent<Transform>()->SetScale(Vector3(5.0f, 5.0f, 5.0f));
+		//	musketeerman->SetName(L"Musketeerman");
+		//	musketeerman->SetPlayerObject(player);
+		//	MeshRenderer* musketeermanmr = musketeerman->AddComponent<MeshRenderer>();
+		//	musketeermanmr->SetMesh(Resources::Find<Mesh>(L"CubeMesh"));
+		//	musketeermanmr->SetMaterial(Resources::Find<Material>(L"BasicMaterial"), 0);
+		//	Collider2D* musketeermancol = musketeerman->AddComponent <Collider2D>();
+		//	musketeermancol->SetType(eColliderType::Box);
+		//	musketeermancol->SetSize(Vector3(1.0, 2.0f, 2.0f));
+		//	Rigidbody* spearmanRigidbody = musketeerman->AddComponent<Rigidbody>();
+		//	spearmanRigidbody->SetGround(false);
 
-			camScript->SetLockOnTarget(musketeerman);
-		}
+		//	camScript->SetLockOnTarget(musketeerman);
+		//}
 		
+		//{
+		//	Swordsman* swordsman = object::Instantiate<Swordsman>(eLayerType::Monster);
+		//	swordsman->GetComponent<Transform>()->SetPosition(Vector3(5.0f, 0.0f, 15.0f));
+		//	swordsman->GetComponent<Transform>()->SetScale(Vector3(5.0f, 5.0f, 5.0f));
+		//	swordsman->SetName(L"Swordsman");
+		//	swordsman->SetPlayerObject(player);
+		//	MeshRenderer* swordsmanmr = swordsman->AddComponent<MeshRenderer>();
+		//	swordsmanmr->SetMesh(Resources::Find<Mesh>(L"CubeMesh"));
+		//	swordsmanmr->SetMaterial(Resources::Find<Material>(L"BasicMaterial"), 0);
+		//	Collider2D* swordsmancol = swordsman->AddComponent <Collider2D>();
+		//	swordsmancol->SetType(eColliderType::Box);
+		//	swordsmancol->SetSize(Vector3(1.0, 2.0f, 2.0f));
+		//	Rigidbody* spearmanRigidbody = swordsman->AddComponent<Rigidbody>();
+		//	spearmanRigidbody->SetGround(false);
+		//	swordsman->AddComponent<MonsterScript>();
+
+		//	camScript->SetLockOnTarget(swordsman);
+		//}
+
 
 		{
 			GameObject* ground = object::Instantiate<GameObject>(eLayerType::Ground);
 			ground->SetName(L"Ground");
 			Transform* groundTr = ground->GetComponent<Transform>();
 			groundTr->SetPosition(Vector3(0.0f, -10.0f, 10.0f));
-			groundTr->SetScale(Vector3(100.0f, 4.0f, 100.0f));
+			groundTr->SetScale(Vector3(1000.0f, 4.0f, 1000.0f));
 			groundTr->SetRotation(Vector3(0.0f, 0.0f, 0.0f));
 			MeshRenderer* groundRenderer = ground->AddComponent<MeshRenderer>();
 			groundRenderer->SetMesh(Resources::Find<Mesh>(L"CubeMesh"));
@@ -123,7 +144,7 @@ namespace ya
 			ground->AddComponent<GroundScript>();
 		}
 
-		{
+		/*{
 			GameObject* grappleGround = object::Instantiate<GameObject>(eLayerType::Ground);
 			grappleGround->SetName(L"grapple target");
 			Transform* groundTr = grappleGround->GetComponent<Transform>();
@@ -152,23 +173,23 @@ namespace ya
 				hookTargetCollider->SetSize(Vector3(1.0, 1.0f, 1.0f));
 				hookTarget->AddComponent<HookTargetScript>();
 			}
-		}
+		}*/
 
-		{
-				GameObject* logbridge = object::Instantiate<GameObject>(eLayerType::Logbridge);
-				logbridge->SetName(L"LogBridge");
-				Transform* logbridgeTr = logbridge->GetComponent<Transform>();
-				logbridgeTr->SetPosition(Vector3(0.0f, -6.0f, 10.0f));
-				logbridgeTr->SetScale(Vector3(10.0f, 2.0f, 50.0f));
-				logbridgeTr->SetRotation(Vector3(0.0f, 0.0f, 0.0f));
-				MeshRenderer* logbridgeRenderer = logbridge->AddComponent<MeshRenderer>();
-				logbridgeRenderer->SetMesh(Resources::Find<Mesh>(L"CubeMesh"));
-				logbridgeRenderer->SetMaterial(Resources::Find<Material>(L"BasicMaterial"),0);
-				Collider2D* logbridgeCollider = logbridge->AddComponent<Collider2D>();
-				logbridgeCollider->SetType(eColliderType::Box);
-				logbridgeCollider->SetSize(Vector3(1.0, 1.0f, 1.0f));
-				logbridge->AddComponent<LogBridgeScript>();
-		}
+		//{
+		//		GameObject* logbridge = object::Instantiate<GameObject>(eLayerType::Logbridge);
+		//		logbridge->SetName(L"LogBridge");
+		//		Transform* logbridgeTr = logbridge->GetComponent<Transform>();
+		//		logbridgeTr->SetPosition(Vector3(0.0f, -6.0f, 10.0f));
+		//		logbridgeTr->SetScale(Vector3(10.0f, 2.0f, 50.0f));
+		//		logbridgeTr->SetRotation(Vector3(0.0f, 0.0f, 0.0f));
+		//		MeshRenderer* logbridgeRenderer = logbridge->AddComponent<MeshRenderer>();
+		//		logbridgeRenderer->SetMesh(Resources::Find<Mesh>(L"CubeMesh"));
+		//		logbridgeRenderer->SetMaterial(Resources::Find<Material>(L"BasicMaterial"),0);
+		//		Collider2D* logbridgeCollider = logbridge->AddComponent<Collider2D>();
+		//		logbridgeCollider->SetType(eColliderType::Box);
+		//		logbridgeCollider->SetSize(Vector3(1.0, 1.0f, 1.0f));
+		//		logbridge->AddComponent<LogBridgeScript>();
+		//}
 
 		CollisionManager::CollisionLayerCheck(eLayerType::Player, eLayerType::Player, true);
 		CollisionManager::CollisionLayerCheck(eLayerType::Player, eLayerType::Monster, true);
